Pal tracker Rest

+++ src/test/java/test/pivotal/pal/trackerapi/TimeEntryApiTest.java
@@ -125,6 +125,6 @@ public class TimeEntryApiTest {
 
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
 
-        return response.getBody().getId();
+        return (Long) response.getBody().getId();
     }
 }
